Remove LD_LIBRARY_PATH usage in tests
