+++ source/neuropods/tests/test_default_backend.cc
@@ -4,16 +4,6 @@
 
 #include "test_utils.hh"
 
-// We need to override the default backend lookup paths
-// Modifying the RPATH doesn't work on Linux because Bazel symlinks all the objects
-// into the test directory
-const std::unordered_map<std::string, std::string> default_backend_overrides = {
-    {"tensorflow", "./neuropods/backends/tensorflow/libneuropod_tensorflow_backend.so"},
-    {"python", "./neuropods/backends/python_bridge/libneuropod_pythonbridge_backend.so"},
-    {"pytorch", "./neuropods/backends/python_bridge/libneuropod_pythonbridge_backend.so"},
-    {"torchscript", "./neuropods/backends/torchscript/libneuropod_torchscript_backend.so"},
-};
-
 TEST(test_default_backend, test_pytorch_addition_model)
 {
     // Test the PyTorch addition model
